Adds user policy

+++ app/controllers/api/v1/users_controller.rb
@@ -1,13 +1,18 @@
 class Api::V1::UsersController < Api::V1::BaseController
+  after_action :verify_authorized
+
   def index
+    authorize User
     render_collection(User, ::V1::UserSerializer)
   end
 
   def show
+    authorize user
     render json: user, status: 200, serializer: ::V1::UserSerializer
   end
 
   def create
+    authorize User
     user = User.new(user_params)
     if user.save
       render json: user, status: 201, serializer: ::V1::UserSerializer
@@ -17,6 +22,7 @@ class Api::V1::UsersController < Api::V1::BaseController
   end
 
   def update
+    authorize user
     if user.update(user_params)
       render json: user, status: 200, serializer: ::V1::UserSerializer
     else
@@ -25,6 +31,7 @@ class Api::V1::UsersController < Api::V1::BaseController
   end
 
   def destroy
+    authorize user
     user.destroy
     head 204
   end
